Rename examples select component to kebab-case
