+++ packages/examples/src/components/examples/SettingsExample/AccountTab.tsx
@@ -14,7 +14,7 @@ import {
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from '@/components/ui/Select';
+} from '@/components/ui/select';
 
 export const AccountTab = forwardRef<
   HTMLDivElement,
