+++ packages/examples/src/components/examples/SettingsExample/PrivacyTab.tsx
@@ -13,7 +13,7 @@ import {
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from '@/components/ui/Select';
+} from '@/components/ui/select';
 
 export const PrivacyTab = forwardRef<
   HTMLDivElement,
