--- packages/examples/src/components/ui/Select/index.ts
@@ -1,9 +1,0 @@
-export { Select, SelectGroup, SelectValue } from '@radix-ui/react-select';
-
-export * from './SelectContent';
-export * from './SelectItem';
-export * from './SelectLabel';
-export * from './SelectScrollDownButton';
-export * from './SelectScrollUpButton';
-export * from './SelectSeparator';
-export * from './SelectTrigger';
