+++ packages/examples/src/components/ui/select/index.ts
@@ -1,0 +1,9 @@
+export { Select, SelectGroup, SelectValue } from '@radix-ui/react-select';
+
+export * from './select-content';
+export * from './select-item';
+export * from './select-label';
+export * from './select-scroll-down-button';
+export * from './select-scroll-up-button';
+export * from './select-separator';
+export * from './select-trigger';
