+++ packages/examples/src/components/ui/select/select-trigger.tsx
